+++ demo-pwa/src/global/app.ts
@@ -1,0 +1,11 @@
+import '@ionic/core';
+import 'ionicons';
+import 'web-animations-js';
+import '@proyecto26/animatable-component';
+import { setupConfig } from '@ionic/core';
+
+export default () => {
+  setupConfig({
+    mode: 'ios'
+  });
+};